Added GameService
Added the game service systems, it is the system that will handle every service of the game like the puzzle manager who became the shardService (name subject to change). This service provide us with every puzzling and system to be centralized and handled by one parent script, the GameInitializer.
Also added a prefab game object to update the shard and interactable object when player enter a room